Implement preview command which renders a component in a test canvas for development.
