+++ package.json
@@ -42,7 +42,6 @@
     "nodemon": "^1.9.1",
     "phantomjs-prebuilt": "^2.1.7",
     "react": "^0.14.7",
-    "react-inlinesvg": "^0.4.2",
     "webpack": "^1.12.14",
     "webpack-dev-server": "^1.14.1"
   }
