+++ src/preview-server/views/index.ejs
@@ -1,19 +1,20 @@
 <!doctype html>
 <html>
-    <head>
-      <meta charset="utf-8">
-      <meta http-equiv="x-ua-compatible" content="ie=edge">
-      <meta name="description" content="">
-      <meta name="viewport" content="width=device-width, initial-scale=1">
-      <title>Sky Component Preview</title>
-      <link href="//assets.sky.com/dist/1268/css/fonts.css" rel="stylesheet">
-      <link href="//assets.sky.com/dist/1268/fonts/skyicons.css" rel="stylesheet">
-      <link rel="stylesheet" href="/component/<%= brand %>.css">
-    </head>
-    <body style="margin: 0;">
+<head>
+  <meta charset="utf-8">
+  <meta http-equiv="x-ua-compatible" content="ie=edge">
+  <meta name="description" content="">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <title>Sky Component Preview</title>
+  <link href="//assets.sky.com/dist/1268/css/fonts.css" rel="stylesheet">
+  <link href="//assets.sky.com/dist/1268/fonts/skyicons.css" rel="stylesheet">
+  <link rel="stylesheet" href="/assets/bundle.css">
+</head>
+<body>
 
-      <%- component %>
-      <script src="/component/<%= brand %>.js"></script>
+  <div id="container"></div>
 
-    </body>
+  <script src="/assets/bundle.js"></script>
+
+</body>
 </html>
